+++ package.json
@@ -22,6 +22,7 @@
         "@fortawesome/react-fontawesome": "^0.1.9",
         "@zeit/next-css": "^1.0.1",
         "@zeit/next-sass": "^1.0.1",
+        "awilix": "^6.1.0",
         "axios": "^0.19.2",
         "body-parser": "^1.19.0",
         "compression": "^1.7.4",
